frontend/requirements.txt
streamlit==1.55.0
pandas==2.3.3
plotly==6.6.0
requests==2.32.5
python-dotenv==1.2.2
numpy==2.4.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
altair==6.0.0
pydeck==0.9.1
pillow==12.1.1
pyarrow==23.0.1
watchdog==6.0.0
blinker==1.9.0
cachetools==7.0.2
packaging==26.0
click==8.3.1
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3